[FEAT] - Add Score and rename some Function

+++ app/src/main/java/com/example/lepeenice/MemoryClassPackage/Sword.kt
@@ -1,7 +1,7 @@
 package com.example.lepeenice.MemoryClassPackage
 
 @kotlinx.serialization.Serializable
-open class Sword(var name: String, var damage: Float) {
+open class Sword(var name: String, var damage: Int) {
 
     override fun toString(): String {
         return "Sword(name=$name, damage=$damage)"
